### 2021-12-09 #3985 @whedon
[PRE REVIEW]: SurfinPy 2.0:  A Phase Diagram Generator for Surfaces and Bulk Phases
**Submitting author:** @jstse (<a href="http://orcid.org/0000-0002-1320-557X">Joshua Tse</a>)
**Repository:** <a href="https://github.com/symmy596/SurfinPy" target ="_blank">https://github.com/symmy596/SurfinPy</a>
**Version:** v2.0
