### 2021-12-09 #3985 @danielskatz
@jstse - your paper does not have any references, which is very unusual. Is there no previous work that you are basing this work on, such as papers, books, software, etc.?  Please see https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography for an example of how to do this.
