### 2021-12-10 #3985 @danielskatz
Also a note from the submission for @lucydot (or the editor, if someone else)
> surfinpy v2 is not considered for publication elsewhere, and adds to surfinpy v1 ([JOSS, DOI 10.21105/joss.01210](https://doi.org/10.21105/joss.01210)). v2 is a substantial addition to v1, which was limited to surface phase diagrams. v2 allows solid state chemists and physicists to build phase diagrams for bulk phases using data from DFT calculations including entropic terms from vibrational frequencies, and where available experimental corrections. v2's publication in JOSS will precede research papers from the Huddersfield and Bath research groups on a wide range of materials, including the MgO-CO2-H2O system, the doping of SrTiO3, and hydrogen defects in UO2.
