### 2021-12-14 #3985 @lucydot
Hi @jstse I'll be your editor for this submission 👋

@dandavies99 are you available to review this submission? Note you may need to check conflict of interest - the JOSS CoI policy is [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy)

@mkhorton are you available to review again for JOSS?
