### 2021-12-16 #3985 @lucydot
Thanks @dandavies99 - I'll assign you now.

And thanks @mkhorton I can see you are a very active reviewer for JOSS  ✨  - so can particularly understand why you may not be available to review for us now. I'll ping a few other people to see if they can help and take it from there.
