### 2021-12-16 #3985 @lucydot
@awvwgk, @amandadumi or @jeffhammond are you available to review this submission? It appears that each of you have the Python and domain knowledge needed.
