### 2021-12-16 symmy596/SurfinPy#20 @awvwgk
Coverage and coveralls in install requirements
Is there a reason to require [coverage](https://coverage.readthedocs.io/en/6.2/) and [coveralls](https://coveralls-python.readthedocs.io/en/latest/) for installation of this project? Those seem more suitable as test requirements for the continuous integration rather than a user installation.
