### 2021-12-16 symmy596/SurfinPy#21 @awvwgk
License file is not included in sdist
The license file is currently not included in the sdist provided on PyPI.
