### 2021-12-18 symmy596/SurfinPy#22 @awvwgk
Jupyter in install requirements
The install requirements in setup.py list Jupyter, however this is not listed in the README.
