### 2021-12-20 #3985 @lucydot
@awvwgk @dandavies99 @jstse it looks like we are ready to start the review! ✨

I'll now ask our bot Whedon to start a review issue thread (this is the pre-review thread).

Please email me or (preferred) ping me in the new review thread with any questions you have about the process. There are review guidelines and a checklist at the top of the new thread (the link will appear below), and there is further reviewer documentation [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

We're normally asking for reviews to be complete within a 4-6 week period however I think given that many of us will be taking a break over Christmas (myself included) and so a 1-2 week delay might be expected and is reasonable. As the review is usually an iterative process it's best for reviews to start earlier in that timeline. You don't need to write a full review in one sitting - rather it's a back-and-forth conversation between reviewers, author, editor and our bot Whedon.
