### 2021-12-20 #4014 @whedon
[REVIEW]: SurfinPy 2.0:  A Phase Diagram Generator for Surfaces and Bulk Phases
**Submitting author:** @jstse (<a href="http://orcid.org/0000-0002-1320-557X">Joshua Tse</a>)
**Repository:** https://github.com/symmy596/SurfinPy
**Branch with paper.md** (empty if default branch):
**Version:** v2.0.2

@awvwgk @dandavies99, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:

  *Note: license is not included in source distribution on PyPI (https://github.com/symmy596/SurfinPy/issues/21)*

  *Note: project dependencies are slightly incorrect, mostly overdepending (https://github.com/symmy596/SurfinPy/issues/22, https://github.com/symmy596/SurfinPy/issues/20)*
