### 2021-12-21 symmy596/SurfinPy#24 @dandavies99
Discrepancies in examples
Related to [JOSS review](https://github.com/openjournals/joss-reviews/issues/4014)

The examples are excellent - really clear and easy to follow. I encountered just a few problems:

- `mu_vs_mu.calculate()`  returns a tuple which breaks Surfaces Tutorials 1, 2 and 3 (and the docstring needs updating accordingly).

- I seem to get a different answer for bulk/Tutorial_4 compared to the plots that are pre-loaded in the notebook. See below for the first example. I'm not sure what the source of the difference could be but might be something you'd like to check.

- Similar situation for the last plot of bulk/Tutorial_6
