### 2021-12-21 symmy596/SurfinPy#23 @awvwgk
> * `pip install surfinpy` results in a successful installation, but `python setup.py test` fails due to `no module named 'yaml'`. Suggest that `pyyaml` needs adding to `install_requires` and on the version on PyPI.

Interesting, didn't notice this when installing in a conda env and `pip check` didn't find an issue as well.
