### 2021-12-21 symmy596/SurfinPy#23 @dandavies99
Yeah, that makes sense. Over in [surfaxe](https://github.com/SMTG-UCL/surfaxe) we switched to `ruaml` (e.g. `from ruamel.yaml import YAML`) the same as pymatgen, seeing as it was a dependency anyway.
