### 2021-12-21 #4014 @dandavies99
Hi @jstse - The code, docs and examples are looking really good to me. I've just got a few points for improvement:

## Installation
I think the installation instructions probably need another pass, largely because of complications caused by pymatgen versions. I opened a [separate issue](https://github.com/symmy596/SurfinPy/issues/23) for this because it may need a bit of extra discussion.

## Functionality
The functionality of the code is as advertised - there are a few minor points which came to light while going through the (excellent) tutorials - I've opened another [separate issue](https://github.com/symmy596/SurfinPy/issues/24).

## Documentation
...is very clear! Might be worth providing an obvious link to the readthedocs hosted version in the README as it wasn't initially obvious to me that these were available - the badge is easy to miss.

## Software paper
The paper looks good and ticks most boxes. Just a few suggestions:

- My main concern is that there is a fair bit of repetition, so it could be made more concise and therefore accessible. For example, it is mentioned ~five times that this release expands functionality to bulk phases. Possibly the numbered list of new features is the best place to keep this, and the following paragraph could simply highlight the temperature range improvement? It probably doesn't need repeating again in the Statement of Need.

- Any mention of the state of the field is missing. For example, you probably need to include how the new functionality for bulk phases is different to what is in the pymatgen.analysis.phase_diagram [module](https://pymatgen.org/pymatgen.analysis.phase_diagram.html), which has a `PhaseDiagram` class with lots of methods to do similar things to this. ASE also has [some phase diagram capabilities](https://wiki.fysik.dtu.dk/ase/ase/phasediagram/phasediagram.html).

- Related to the above, probably a couple of citations needed beyond self-citations. These might be to existing software (pymatgen, ASE) but also a key reference or two to the main theory or theories that SurfinPy is using, if you think that's appropriate.

- Some other minor suggestions from me but perhaps these are more editorial considerations for @lucydot:
	- Para 1: "can and have been used" -> "have been used"
	- "This method uses the vibrational modes for each structure to calculate the zero point energy and thus the vibrational entropy." - it's a bit unclear what "each structure" refers to.
	- "and applies a higher level of theory"- is a bit vague? Maybe you don't need this.
