### 2021-12-21 symmy596/SurfinPy#25 @awvwgk
PyPI releases mismatch tags?
What is the relation between the releases/tags in this repo (https://github.com/symmy596/SurfinPy/tags) and the releases available on PyPI (https://pypi.org/project/surfinpy/#history). There are several versions appearing on PyPI which are not tagged in this repository.
