### 2022-01-01 #4014 @lucydot
Hi @dandavies99 thank-you for your quick and timely review 👍

Hi @jstse - RE: the paper. I agree with the suggestions @dandavies99 makes above, in addition:
- "in order to improve performance": this is a little vague. Is this a speed-up compared to the previous software release(s)?
- "Similarly, to the surface module,": the wording here doesn't sound right: perhaps re-phrased to "Similar to the the surface module". Although, re-iterating Dan's point, you may want to do a larger re-structuring of this section to avoid repetition.

State of the field, and how your package compares to other packages with an overlap of functionality (pymatgen/ASE) is important. I'll include the relevant quote from the [author guidelines](https://joss.readthedocs.io/en/latest/submitting.html):

> Your software should be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler).

If you have any questions or if I can clarify anything, please let me know.
