### 2022-01-03 b233de1
Fix typos in documentation
- automatically remove trailing whitespace
