+++ docs/source/tutorial_1.rst
@@ -1,15 +1,15 @@
 Chemical Potential
 ==================
 
-The physical quantity that is used to define the stability of a surface with a given composition is its surface energy :math:`\gamma` (J :math:`m^{-2}`). 
+The physical quantity that is used to define the stability of a surface with a given composition is its surface energy :math:`\gamma` (J :math:`m^{-2}`).
 At its core, `surfinpy` is a code that calculates the surface energy of different slabs and uses these surface energies to build a phase diagram.
-In this explantion of theory we will use the example of water adsorbing onto a surface of :math:`TiO_2` containing oxygen vacancies.
-In such an example there are two variables, water concentration and oxygen vacancy concentration. We are able to calculate the surface energy according to 
+In this explanation of theory we will use the example of water adsorbing onto a surface of :math:`TiO_2` containing oxygen vacancies.
+In such an example there are two variables, water concentration and oxygen vacancy concentration. We are able to calculate the surface energy according to
 
 .. math::
     \gamma_{Surf} = \frac{1}{2A} \Bigg( E_{TiO_2}^{slab} - \frac{nTi_{slab}}{nTi_{Bulk}} E_{TiO_2}^{Bulk} \Bigg) - \Gamma_O \mu_O - \Gamma_{H_2O} \mu_{H_2O} ,
 
-where A is the surface area, :math:`E_{TiO_2}^{slab}` is the DFT energy of the slab, :math:`nTi_{Slab}` is the number of cations in the slab, 
+where A is the surface area, :math:`E_{TiO_2}^{slab}` is the DFT energy of the slab, :math:`nTi_{Slab}` is the number of cations in the slab,
 :math:`nTi_{Bulk}` is the number of cations in the bulk unit cell, :math:`E_{TiO_2}^{Bulk}` is the DFT energy of the bulk unit cell and
 
 .. math::
@@ -18,28 +18,28 @@ where A is the surface area, :math:`E_{TiO_2}^{slab}` is the DFT energy of the s
 .. math::
     \Gamma_{H_2O} = \frac{nH_2O}{2A} ,
 
-where :math:`nO_{Slab}` is the number of anions in the slab, :math:`nO_{Bulk}` is the number of anions in the bulk and :math:`nH_2O` is the number of adsorbing water molecules. 
-:math:`\Gamma_O` / :math:`\Gamma_{H_2O}` is the excess oxygen / water at the surface and :math:`\mu_O` / :math:`\mu_{H_2O}` is the oxygen / water chemcial potential. 
-Clearly :math:`\Gamma` and :math:`mu` will only matter when the surface is non stoichiometric. 
+where :math:`nO_{Slab}` is the number of anions in the slab, :math:`nO_{Bulk}` is the number of anions in the bulk and :math:`nH_2O` is the number of adsorbing water molecules.
+:math:`\Gamma_O` / :math:`\Gamma_{H_2O}` is the excess oxygen / water at the surface and :math:`\mu_O` / :math:`\mu_{H_2O}` is the oxygen / water chemical potential.
+Clearly :math:`\Gamma` and :math:`mu` will only matter when the surface is non stoichiometric.
 
 Usage
 ~~~~~
 
-The first thing to do is input the data that we have generated from our DFT calculations. 
+The first thing to do is input the data that we have generated from our DFT calculations.
 The input data needs to be contained within a dictionary.
[…]
