### 2022-01-03 #4014 @awvwgk
### Installation

- the license file is currently not included with in the main distribution channel (PyPI)
  - https://github.com/symmy596/SurfinPy/issues/21
- the package has more dependencies than needed for the user
  - https://github.com/symmy596/SurfinPy/issues/20
  - https://github.com/symmy596/SurfinPy/issues/22
- the main distribution channel lists different version compared to the repo
  - https://github.com/symmy596/SurfinPy/issues/25
- installation instructions could use overhaul (already mentioned by @dandavies99)
  - `pip install .` instead of calling `setup.py`
  - recommend usage of venv / conda

### Documentation

Generally, in a really good shape

- lots of typos and duplication (noticed the duplicates because of identical typos in tutorials and theory)
  - https://github.com/symmy596/SurfinPy/pull/26
- documentation should be clearly linked in the README / repository description

### Software paper

- state of the field should be discussed in more details
- is Vasp the only available backend for this project? What about other (free) electronic structure programs?
