### 2022-01-10 #4014 @lucydot
Hello @jstse, pinging you just in case you have missed the feedback from @awvwgk above ☝️.

It looks like you have a few items to be working on, please also let us know how your progress is going here in the thread.
