+++ paper/paper.md
@@ -31,44 +31,37 @@ bibliography: paper.bib
 
 # Summary
 
-SurfinPy is a python module for generating phase diagrams from Density Functional Theory data. 
-The previous code release, reported in Adam R. Symington *et. al.* '[@Symington2019]', calculated the surface free energy under different external conditions, and used these surface free energies to generate phase diagrams. 
-These phase diagrams can and have been used to provide an understanding of surface composition under various environmental conditions, thus providing crucial information for a range of materials science problems '[@Symington2020c;@Symington2020a;@Moxon2020]'. 
-In this new release, the code has been expanded to generate phase diagrams for bulk phases, as well as surface phases. 
+SurfinPy is a python module for generating phase diagrams from data derived from ab initio and/or classical techniques.
 
-This `SurfinPy` release makes the following contributions
+The previous code release, reported in Adam R. Symington *et. al.* , calculated the surface free energy under different external conditions, and used surface free energy values to generate phase diagrams. 
+Phase diagrams have been used to provide an understanding of surface composition under various environmental conditions, thus providing crucial information for a range of materials science problems '[@Symington2020c;@Symington2020a;@Moxon2020]'. 
+In this second release, the code has been expanded to generate phase diagrams for bulk phases, as well as surface phases. 
 
-1. Introduces a new capability to calculate phase diagrams for bulk phases as a function of the chemical potential (and/or pressure) of two species.
-2. Introduces a new capability to calculate phase diagrams for bulk phases as a function of the chemical potential and temperature.
-3. Allows the calculation of vibrational entropy and zero point energy for solid phases and introduce these values into the phase diagram generation.
-4. A significant update to the original code base in order to improve performance, streamline workflow and improve the plotting options.
+This `SurfinPy` release increases the capability of the previous release implementing the following major contributions
 
-The largest contribution to this release is the ability to calculate and plot phase diagrams for bulk phases, which is a substantial expansion of the previous code to include free
-energy for bulk phases (and not just surfaces) under various pressures and temperatures.
-Similarly, to the surface module, free energies for each phase are calculated under a specific temperature and pressure, which is then used to generate phase diagrams for bulk phases. 
+1.	The calculation of phase diagrams for bulk phases.
+2. A significant update to the original code base in order to improve performance both in terms of capability and speedup, stream-line workflow and improve the plotting options.
+
+The largest contribution to this release is the ability to calculate free energy of bulk phases (and not just surface phases) under various pressures and temperatures and plot phase diagrams for bulk phases as a function  of the chemical potential (and/or pressure) of two species, and as a function  of the chemical potential and temperature. 
+Similarly to the surface module, in the bulk module, free energies of different phases are calculated under specific temperatures and pressures, which are then used to generate phase diagrams for bulk phases. 
 Instead of an absolute value, temperature ranges can now be provided, enabling the ability to plot pressure as a function of temperature (or vice-versa) providing results, which are comparable to experimental data where available.
 
-Another notable addition to this release is the ability to calculate vibrational properties for solid phases from Density Functional Theory calculations. 
-This method uses the vibrational modes for each structure to calculate the zero point energy and thus the vibrational entropy. 
-These values can then be used to improve the accuracy of the free energy calculation and applies a higher level of theory. An example of bulk phase diagrams is shown in Figure 1.
-
-Finally, eleven tutorials have been developed to highlight the full functionality of this new `SurfinPy` release. These are all available in jupyter notebooks in the repository. 
+Another notable addition to this release is the ability to calculate vibrational properties for solid bulk phases. 
+The vibrational modes for each solid bulk phase are used to calculate the zero point energy and thus the vibrational entropy. 
[…]
