+++ paper/paper.md
@@ -61,7 +61,7 @@ Finally, eleven tutorials have been developed to highlight the full functionalit
 With this release SurfinPy is no longer limited to the surface composition and chemistry and particle morphology, but expands on the chemistry of bulk phases, which make it ideal for applications to the broad spectrum of research questions around solid state chemistry and physics. 
 This allows the exploration of the phase stability of solid state systems (bulk and surface phases) of different compositions as a function of external conditions. 
 
-Other codes capable to produce phase diagrams *e.g.* pymatgen '[ONG2013314]' and ASE '[Hjorth_Larsen_2017]' are available.  However, our code is self-contained allowing for the generating both bulk and surface phase diagrams, while offering easier and enhanced plotting capability to compare phases as a function of chemical potential of different species and temperature.  Additionally, unlike other codes detailed tutorials are available, offering a more tailored and focused experience compared to other codes.
+Other codes capable to produce phase diagrams *e.g.* pymatgen `[ONG2013314]` and ASE `[Hjorth_Larsen_2017]` are available.  However, our code is self-contained allowing for the generating both bulk and surface phase diagrams, while offering easier and enhanced plotting capability to compare phases as a function of chemical potential of different species and temperature.  Additionally, unlike other codes detailed tutorials are available, offering a more tailored and focused experience compared to other codes.
 
 # Acknowledgements
   
