### 2022-01-27 2b10910
Updates to Tutorial and Paper
