+++ paper/paper.md
@@ -31,37 +31,31 @@ bibliography: paper.bib
 
 # Summary
 
-SurfinPy is a python module for generating phase diagrams from data derived from ab initio and/or classical techniques.
+`SurfinPy` is a python module for generating phase diagrams from data derived from ab initio and/or classical methodologies.
 
-The previous code release, reported in Adam R. Symington *et. al.* [@Symington2019], calculated the surface free energy under different external conditions, and used surface free energy values to generate phase diagrams. 
-Phase diagrams have been used to provide an understanding of surface composition under various environmental conditions, thus providing crucial information for a range of materials science problems [@Symington2020c;@Symington2020a;@Moxon2020]. 
-In this second release, the code has been expanded to generate phase diagrams for bulk phases, as well as surface phases. 
+The previous code release, reported in Adam R. Symington *et. al.* [@Symington2019], calculated the surface free energy values under different external conditions, and used these values to generate phase diagrams. 
+Surface phase diagrams have been used to provide an understanding of surface composition under various environmental conditions, thus giving crucial information for a range of surface science problems [@Symington2020c;@Symington2020a;@Moxon2020]. 
 
-This `SurfinPy` release increases the capability of the previous release implementing the following major contributions
+In this second `SurfinPy` release, the capability of the code has been expanded to generate phase diagrams for bulk phases, as well as surface phases. The code how now the ability to calculate free energy values of bulk phases under specific values of pressure and temperature, and use these to plot phase diagrams for bulk phases as a function  of the chemical potential (and/or pressure) of two species, and as a function  of the chemical potential and temperature. 
+Instead of an absolute value, temperature ranges can now be provided, enabling the ability to plot pressure as a function of temperature (or vice-versa) giving results, that are comparable to experimental data where available.
+Another notable addition to this release is the ability to calculate vibrational properties for bulk phases. 
+The vibrational modes for bulk phases are used to calculate the zero point energy and thus the vibrational entropy. 
+The code allows for the inclusion of these values into the generation of phase diagrams, which removes the approximation that entropy of bulk phases has little contribution to the free energy and may improve the accuracy of the methodology.
 
-1.	The calculation of phase diagrams for bulk phases.
-2. A significant update to the original code base in order to improve performance both in terms of capability and speedup, stream-line workflow and improve the plotting options.
-
-The largest contribution to this release is the ability to calculate free energy of bulk phases (and not just surface phases) under various pressures and temperatures and plot phase diagrams for bulk phases as a function  of the chemical potential (and/or pressure) of two species, and as a function  of the chemical potential and temperature. 
-Similarly to the surface module, in the bulk module, free energies of different phases are calculated under specific temperatures and pressures, which are then used to generate phase diagrams for bulk phases. 
-Instead of an absolute value, temperature ranges can now be provided, enabling the ability to plot pressure as a function of temperature (or vice-versa) providing results, which are comparable to experimental data where available.
-
-Another notable addition to this release is the ability to calculate vibrational properties for solid bulk phases. 
-The vibrational modes for each solid bulk phase are used to calculate the zero point energy and thus the vibrational entropy. 
-The code allows for the inclusion of these values into the generation of phase diagrams. 
-This removes the approximation that entropy of solid phases does not contribute to the free energy and may improve the accuracy of the methodology.
+A significant update to the original code has also been made to improve performance in terms of speedup, stream-line workflow and enhanced plotting options.
 
 ![An example of a phase diagram as a function of chemical potential.\label{fig:example}](surfinpy.pdf)
+
 Finally, eleven tutorials have been developed to highlight the full functionality of this new SurfinPy release. These are all available in jupyter notebooks in the repository.
[…]
