### 2022-02-07 symmy596/SurfinPy#22 @jstse
The examples we provide are created using jupyter notebooks and we feel for new user experience having jupyter would be a requirement.
