### 2022-02-07 d5505dd
Update install requirements
- Removed Pandas from readme
- Removed version constraints in install
- Updated SciPy constants to use value over depreciated codata.
