+++ requirements.txt
@@ -1,11 +1,11 @@
-jupyter>=1.0.0
-pymatgen>=2020.1.28
-numpy>=1.17.2
-matplotlib>=3.1.1
-pyyaml>=5.1.2
-seaborn>=0.9.0
-scipy>=1.5.2
-pandas>=0.25.1
-nbsphinx>=0.7.1
-jupyter-sphinx>=0.3.1
-sphinx_rtd_theme>=0.5.0
+jupyter
+pymatgen
+numpy
+matplotlib
+pyyaml
+seaborn
+scipy
+pandas
+nbsphinx
+jupyter-sphinx
+sphinx_rtd_theme
