### 2022-02-07 symmy596/SurfinPy#24 @jstse
Updated tutorail pictures and should now match the plots you recieved.  Issue was old pictures in tutorials not being updated when changing plotting values for examples.  See pull request #27
