### 2022-02-07 symmy596/SurfinPy#23 @jstse
Removed the requirements for specific versions as we do not required specific versions.  Hopefully, this will correct the installation issues you was having.

Updated the local instalation instructions as requested to use `pip install -r requirements.txt -e .`

See Pull request for more information #27
