### 2022-02-07 symmy596/SurfinPy#20 @jstse
Removed coverage and coveralls from installation requirments.  See pull request for more information #27
