### 2022-02-07 symmy596/SurfinPy#22 @awvwgk
Jupyter is not crucial for the functionality of the package, therefore it doesn't qualify as an install requirement. For optional dependencies there is a dedicated mechanism in setup.py which allows installations with features (*i.e.* `pip install .[jupyter]`).
