### 2022-02-07 b250b5d
Moved Jupyter to optional requirement
