+++ setup.py
@@ -24,7 +24,8 @@ if __name__ == "__main__":
         packages=['surfinpy'],
         zip_safe=False,
         python_requires='>=3.6',
-        install_requires=['scipy', 'numpy', 'jupyter', 'pymatgen', 'numpy', 'matplotlib', 'seaborn','pyyaml'],
+        install_requires=['scipy', 'numpy', 'pymatgen', 'numpy', 'matplotlib', 'seaborn','pyyaml'],
+        extras_require={'Tutorials': ["jupyter"]},
         classifiers=['Programming Language :: Python',
                      'Development Status :: 5 - Production/Stable',
                      'Intended Audience :: Science/Research',
