+++ requirements.txt
@@ -1,4 +1,3 @@
-jupyter
 pymatgen
 numpy
 matplotlib
