### 2022-02-07 symmy596/SurfinPy#22 @jstse
Thank you for informing me of this feature.
I have no added jupyter to an optional extra under a Tutorials install and have updated the Readme with install instructions.
