### 2022-02-08 #4014 @jstse
Hi, The pull request has now been merged and will give a detailed response below:

### Installation
- Removed specific versions of requirements causing issues in installation.  After checking became apparent that we do not need specific versions of these codes. (Waiting on confirmation that this has solved installation issues before closing issue)
- As suggested also cleaned up installation instructions
- Removed Coverage and Coveralls from install requirements
- Moved Jupyter to extra requirements and updated insatllation instructions in readme to include insatlling version with Jupyter.
- PyPI mismatch tags, currently waiting for access to the PyPI repository.
-  PyPI License not included, currently waiting for access to the PyPI repository.

### Functionality
- Discrepancies in the tutorials should now be fixed, issues were due to pictures in tutorials not being updated after changing plotting values. (Waiting for confirmation before closing issue)

### Documentation
- As suggested, we have now added a link to the compiled documentation in the Readme.
- Accepted pull request correcting typos in documentation

### Software paper
Dan

> My main concern is that there is a fair bit of repetition, so it could be made more concise and therefore accessible. For example, it is mentioned ~five times that this release expands functionality to bulk phases. Possibly the numbered list of new features is the best place to keep this, and the following paragraph could simply highlight the temperature range improvement? It probably doesn't need repeating again in the Statement of Need.

- Removed list and went into details in the paragraphs after the list of major contributions. This should remove all the repetition. We also have removed the repetition in the statement of need.

`Any mention of the state of the field is missing. For example, you probably need to include how the new functionality for bulk phases is different to what is in the pymatgen.analysis.phase_diagram which has a Phase Diagram class with lots of methods to do similar things to this. ASE also has some phase diagram capabilities.`

- We have now included a sentence acknowledging the other software, and we have highlighted the main advantage of our code. This is added to the statement of need.

`Related to the above, probably a couple of citations needed beyond self-citations. These might be to existing software (pymatgen, ASE) but also a key reference or two to the main theory or theories that SurfinPy is using, if you think that's appropriate.`
- We have accounted the references for the software.

`Para 1: "can and have been used" -> "have been used"`
- Correction accepted.

>  This method uses the vibrational modes for each structure to calculate the zero point energy and thus the vibrational entropy." - it's a bit unclear what "each structure" refers to.

- We have now been consistent with terminology and used "solid bulk phases".

> "and applies a higher level of theory"- is a bit vague? Maybe you don't need this.

- This has now been removed.

Lucy

>  "in order to improve performance": this is a little vague. Is this a speed-up compared to the previous software release(s)?
- We have specified that the improved performance is both in terms of capability and speed up.

> "Similarly, to the surface module,": the wording here doesn't sound right: perhaps re-phrased to "Similar to the the surface module". Although, re-iterating Dan's point, you may want to do a larger re-structuring of this section to avoid repetition.
- correction accepted.

> State of the field, and how your package compares to other packages with an overlap of functionality (pymatgen/ASE) is important. I'll include the relevant quote from the author guidelines:
> Your software should be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler).
- We have now addressed this point in the statement of need.

AWVWGK

> state of the field should be discussed in more details
- We have added a comment in the statement of need as also suggested by the editor and other reviewer.

> is Vasp the only available backend for this project? What about other (free) electronic structure programs?

- No. The code only needs a set of self-contained energies. They can be derived using an ab initio code or a code that implements classical techniques based on force fields. We have indeed added this to the main paper.
