### 2022-02-08 symmy596/SurfinPy#25 @jstse
@symmy596 should now have fixed this issue
