### 2022-02-08 symmy596/SurfinPy#25 @awvwgk
Thanks. Somewhat scary that you just unpublish stuff from PyPI, though.

BTW, I put up a feedstock on cf: https://github.com/conda-forge/surfinpy-feedstock.
