+++ setup.py
@@ -23,7 +23,9 @@ if __name__ == "__main__":
         license='MIT license',
         packages=['surfinpy'],
         zip_safe=False,
-        install_requires=['scipy', 'numpy', 'jupyter', 'pymatgen', 'numpy', 'matplotlib', 'coverage', 'coveralls', 'seaborn'],
+        python_requires='>=3.6',
+        install_requires=['scipy', 'numpy', 'pymatgen', 'numpy', 'matplotlib', 'seaborn','pyyaml'],
+        extras_require={'Tutorials': ["jupyter"]},
         classifiers=['Programming Language :: Python',
                      'Development Status :: 5 - Production/Stable',
                      'Intended Audience :: Science/Research',
