+++ setup.cfg
@@ -1,0 +1,3 @@
+[metadata]
+license = MIT
+license_files = license
