+++ paper/paper.bib
@@ -58,7 +58,7 @@ volume = {68},
 pages = {314-319},
 year = {2013},
 issn = {0927-0256},
-doi = {https://doi.org/10.1016/j.commatsci.2012.10.028},
+doi = {10.1016/j.commatsci.2012.10.028},
 url = {https://www.sciencedirect.com/science/article/pii/S0927025612006295},
 author = {Shyue Ping Ong and William Davidson Richards and Anubhav Jain and Geoffroy Hautier and Michael Kocher and Shreyas Cholia and Dan Gunter and Vincent L. Chevrier and Kristin A. Persson and Gerbrand Ceder},
 }
