### 2022-02-22 #4014 @jstse
Hi @lucydot both of those issues should now be fixed.
