### 2022-02-22 #4014 @lucydot
Excellent @jstse the paper looks good to me.

Please can you do the following:

- Make a tagged release of your software
- Archive this release on Zenodo (or similar)
        - The archive should have exactly the same title and author list as your JOSS paper
- Report the archive DOI and version number in this thread 👇
