+++ docs/source/conf.py
@@ -22,7 +22,7 @@ copyright = '2020, Adam R. Symington'
 author = 'Adam R. Symington, Joshua S. Tse'
 
 # The full version, including alpha/beta/rc tags
-release = '2.0.0'
+release = '2.0.1'
 
 
 # -- General configuration ---------------------------------------------------
