### 2022-03-01 #4014 @lucydot
Hi @symmy596 can you adjust the author list on the Zenodo release so that it matches that on the JOSS paper?
