### 2022-03-01 #4014 @AdamSymington
@lucydot Apologies I completely missed that.
zenodo release with matching authors is now live - https://zenodo.org/record/6250513#.Yh5_BJbLdD8
doi - 10.5281/zenodo.6250513
