### 2022-03-02 #4014 @symmy596
The error is due to me not having an affiliation. The work was completed while I was at the University of Bath so I have altered the paper to include that as my affiliation.
