### 2022-03-03 #4014 @kyleniemeyer
Hi @symmy596, I just submitted a minor PR that fixes some formatting things with the paper: https://github.com/symmy596/SurfinPy/pull/29

Could you merge that? Then, I can accept the paper.
